requirements.txt
numpy
scipy
matplotlib
keras==2.3.1
tensorflow==2.1.0
pathlib2
seaborn
six